refactor: 팀 스크린 파일 구조 리팩토링
