--- src/screens/team/creation/team_creation_style.ts
@@ -1,286 +1,0 @@
-import { StyleSheet } from 'react-native';
-
-import { theme } from '@/src/theme';
-
-export const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: theme.colors.background.main,
-  },
-  inputContainer: {
-    marginBottom: theme.spacing.spacing6,
-  },
-  inputLabel: {
-    fontSize: theme.typography.fontSize.font4,
-    fontWeight: theme.typography.fontWeight.medium,
-    color: theme.colors.text.main,
-    marginBottom: theme.spacing.spacing2,
-  },
-  textInputError: {
-    borderColor: theme.colors.error,
-  },
-  errorText: {
-    fontSize: theme.typography.fontSize.font3,
-    color: theme.colors.error,
-    marginTop: theme.spacing.spacing1,
-  },
-  characterCount: {
-    fontSize: theme.typography.fontSize.font2,
-    color: theme.colors.text.sub,
-    textAlign: 'right',
-    marginTop: theme.spacing.spacing1,
-  },
-  selectorContainer: {
-    marginBottom: theme.spacing.spacing6,
-  },
-
-  stepContainer: {
-    flex: 1,
-    backgroundColor: theme.colors.background.main,
-  },
-  stepHeader: {
-    paddingHorizontal: theme.spacing.spacing5,
-    paddingTop: theme.spacing.spacing10,
-    paddingBottom: 20,
-    alignItems: 'center',
-  },
-  stepTitle: {
-    fontSize: theme.typography.fontSize.font7,
-    fontWeight: theme.typography.fontWeight.bold,
-    color: theme.colors.text.main,
-    textAlign: 'center',
-    marginBottom: theme.spacing.spacing2,
-  },
-  stepSubtitle: {
-    fontSize: theme.typography.fontSize.font4,
-    color: theme.colors.text.sub,
-    textAlign: 'center',
-    lineHeight: 24,
-  },
-  stepContent: {
-    flex: 1,
-    paddingHorizontal: theme.spacing.spacing5,
-    justifyContent: 'center',
-  },
-  stepFooter: {
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    alignItems: 'center',
-    paddingHorizontal: theme.spacing.spacing5,
-    paddingVertical: 20,
-    paddingBottom: theme.spacing.spacing12,
-  },
-  stepTextInput: {
-    borderWidth: 1,
-    borderColor: theme.colors.border.light,
-    borderRadius: theme.spacing.spacing3,
-    paddingHorizontal: theme.spacing.spacing5,
-    paddingVertical: theme.spacing.spacing4,
-    fontSize: theme.typography.fontSize.font5,
-    backgroundColor: theme.colors.background.sub,
-  },
-  stepTextArea: {
-    borderWidth: 1,
-    borderColor: theme.colors.border.light,
-    borderRadius: theme.spacing.spacing3,
-    paddingHorizontal: theme.spacing.spacing5,
-    paddingVertical: theme.spacing.spacing4,
-    fontSize: theme.typography.fontSize.font4,
-    backgroundColor: theme.colors.background.sub,
-    minHeight: 120,
-    textAlignVertical: 'top',
-  },
-  stepSelectorButton: {
-    borderWidth: 1,
-    borderColor: theme.colors.border.light,
-    borderRadius: theme.spacing.spacing3,
-    paddingVertical: theme.spacing.spacing4,
-    paddingHorizontal: theme.spacing.spacing5,
-    marginBottom: theme.spacing.spacing3,
-    backgroundColor: theme.colors.background.sub,
-  },
-  stepSelectorButtonActive: {
-    borderColor: theme.colors.blue[500],
-    backgroundColor: theme.colors.blue[50],
-  },
-  stepSelectorButtonText: {
-    fontSize: theme.typography.fontSize.font4,
-    fontWeight: theme.typography.fontWeight.medium,
-    color: theme.colors.text.sub,
-    textAlign: 'center',
-  },
-  stepSelectorButtonTextActive: {
-    color: theme.colors.blue[500],
-    fontWeight: theme.typography.fontWeight.semibold,
-  },
-  nextButton: {
-    backgroundColor: theme.colors.blue[500],
-    borderRadius: theme.spacing.spacing3,
-    paddingVertical: theme.spacing.spacing3,
-    paddingHorizontal: theme.spacing.spacing5,
-    flexDirection: 'row',
-    alignItems: 'center',
-    justifyContent: 'center',
-    minWidth: theme.spacing.spacing13,
-  },
-  nextButtonDisabled: {
-    backgroundColor: theme.colors.gray[200],
-  },
-  nextButtonText: {
-    color: theme.colors.white,
-    fontSize: theme.typography.fontSize.font4,
-    fontWeight: theme.typography.fontWeight.semibold,
-    marginRight: theme.spacing.spacing2,
-  },
-  backButton: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    justifyContent: 'center',
-    paddingVertical: theme.spacing.spacing4,
-    paddingHorizontal: theme.spacing.spacing8,
-    backgroundColor: theme.colors.background.sub,
-    borderRadius: theme.spacing.spacing3,
-    borderWidth: 1,
-    borderColor: theme.colors.border.light,
-    minWidth: theme.spacing.spacing15,
-  },
-  backButtonText: {
-    color: theme.colors.text.sub,
-    fontSize: theme.typography.fontSize.font4,
-    marginLeft: theme.spacing.spacing1,
-    fontWeight: theme.typography.fontWeight.medium,
-  },
-  memberCountStepContainer: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    justifyContent: 'center',
-    paddingVertical: theme.spacing.spacing10,
-  },
-  memberCountStepButton: {
-    width: theme.spacing.spacing15,
-    height: theme.spacing.spacing15,
-    borderRadius: theme.spacing.spacing15,
-    backgroundColor: theme.colors.background.sub,
-    borderWidth: theme.spacing.spacing1,
-    borderColor: theme.colors.border.light,
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-  memberCountStepDisplay: {
-    alignItems: 'center',
-    marginHorizontal: theme.spacing.spacing10,
-  },
-  memberCountStepNumber: {
-    fontSize: theme.typography.fontSize.font11,
-    fontWeight: theme.typography.fontWeight.bold,
-    color: theme.colors.text.main,
-  },
-  memberCountStepLabel: {
-    fontSize: theme.typography.fontSize.font4,
-    color: theme.colors.text.sub,
-    marginTop: theme.spacing.spacing1,
-  },
-  progressContainer: {
-    paddingVertical: 8,
-    backgroundColor: theme.colors.background.main,
-    marginTop: -theme.spacing.spacing2,
-  },
-  progressBar: {
-    height: theme.spacing.spacing1,
-    backgroundColor: theme.colors.gray[200],
-    borderRadius: theme.spacing.spacing1,
-    marginBottom: theme.spacing.spacing2,
-  },
-  progressFill: {
-    height: '100%',
-    backgroundColor: theme.colors.blue[500],
-    borderRadius: theme.spacing.spacing1,
-  },
-  progressText: {
-    fontSize: theme.typography.fontSize.font2,
-    color: theme.colors.text.sub,
-    textAlign: 'center',
-  },
-  footerSpacer: {
-    flex: 1,
-  },
-  dropdownButton: {
-    borderWidth: 1,
-    borderColor: theme.colors.border.light,
-    borderRadius: theme.spacing.spacing3,
-    paddingHorizontal: theme.spacing.spacing5,
-    paddingVertical: theme.spacing.spacing4,
-    backgroundColor: theme.colors.background.sub,
-    flexDirection: 'row',
-    alignItems: 'center',
-    justifyContent: 'space-between',
-  },
-  dropdownButtonText: {
-    fontSize: theme.typography.fontSize.font5,
-    color: theme.colors.text.main,
-    flex: 1,
-  },
-  placeholderText: {
-    color: theme.colors.text.light,
-  },
-  modalOverlay: {
-    flex: 1,
-    backgroundColor: theme.colors.overlay.dark,
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-  dropdownModal: {
-    backgroundColor: theme.colors.background.main,
-    borderRadius: theme.spacing.spacing4,
-    marginHorizontal: theme.spacing.spacing5,
-    maxHeight: '70%',
-    width: '90%',
-    shadowColor: theme.colors.black,
-    shadowOffset: { width: 0, height: theme.spacing.spacing1 },
-    shadowOpacity: 0.25,
-    shadowRadius: theme.spacing.spacing2,
-    elevation: theme.spacing.spacing2,
-  },
-  dropdownHeader: {
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    alignItems: 'center',
-    paddingHorizontal: theme.spacing.spacing5,
-    paddingVertical: theme.spacing.spacing4,
-    borderBottomWidth: 1,
-    borderBottomColor: theme.colors.border.light,
-  },
-  dropdownTitle: {
-    fontSize: theme.typography.fontSize.font5,
-    fontWeight: theme.typography.fontWeight.semibold,
-    color: theme.colors.text.main,
-  },
-  dropdownItem: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    justifyContent: 'space-between',
-    paddingHorizontal: theme.spacing.spacing5,
-    paddingVertical: theme.spacing.spacing4,
-    borderBottomWidth: 1,
-    borderBottomColor: theme.colors.gray[100],
-  },
-  dropdownItemSelected: {
-    backgroundColor: theme.colors.blue[50],
-  },
-  dropdownItemText: {
-    fontSize: theme.typography.fontSize.font4,
-    color: theme.colors.text.main,
-    flex: 1,
-  },
-  dropdownItemTextSelected: {
-    color: theme.colors.blue[500],
-    fontWeight: theme.typography.fontWeight.medium,
-  },
-  readOnlyField: {
-    backgroundColor: theme.colors.gray[50],
-    borderColor: theme.colors.gray[200],
-  },
-  readOnlyText: {
-    color: theme.colors.text.sub,
-  },
-});
