--- src/screens/team/guide/components/join_waiting_list/styles.ts
@@ -1,382 +1,0 @@
-import { StyleSheet } from 'react-native';
-
-import { colors, spacing, typography } from '@/src/theme';
-
-export const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: colors.white,
-  },
-  header: {
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    alignItems: 'center',
-    paddingHorizontal: spacing.spacing4,
-    paddingVertical: spacing.spacing4,
-    borderBottomWidth: 1,
-    borderBottomColor: colors.gray[200],
-  },
-  title: {
-    fontSize: typography.fontSize.font5,
-    fontWeight: typography.fontWeight.semibold,
-    color: colors.gray[900],
-  },
-  closeButton: {
-    padding: 4,
-    borderRadius: 4,
-  },
-  listContainer: {
-    padding: spacing.spacing4,
-  },
-  joinWaitingItem: {
-    backgroundColor: colors.white,
-    borderRadius: spacing.spacing4,
-    padding: spacing.spacing5,
-    marginBottom: spacing.spacing4,
-    shadowColor: colors.black,
-    shadowOffset: { width: 0, height: 4 },
-    shadowOpacity: 0.1,
-    shadowRadius: 8,
-    elevation: 3,
-    borderWidth: 1,
-    borderColor: colors.gray[100],
-  },
-
-  cardHeader: {
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    alignItems: 'flex-start',
-    marginBottom: spacing.spacing4,
-  },
-  teamInfoSection: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    flex: 1,
-  },
-  teamIconContainer: {
-    width: 48,
-    height: 48,
-    borderRadius: 24,
-    backgroundColor: colors.blue[50],
-    alignItems: 'center',
-    justifyContent: 'center',
-    marginRight: spacing.spacing3,
-    shadowColor: colors.blue[500],
-    shadowOffset: { width: 0, height: 2 },
-    shadowOpacity: 0.1,
-    shadowRadius: 4,
-    elevation: 2,
-  },
-  teamTextContainer: {
-    flex: 1,
-  },
-  teamNameText: {
-    fontSize: typography.fontSize.font5,
-    fontWeight: typography.fontWeight.bold,
-    color: colors.gray[900],
-  },
-  statusBadge: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    paddingHorizontal: spacing.spacing3,
-    paddingVertical: spacing.spacing2,
-    borderRadius: spacing.spacing4,
-    gap: 4,
-  },
-  statusText: {
-    fontSize: typography.fontSize.font2,
-    fontWeight: typography.fontWeight.semibold,
-  },
-
-  applicantSection: {
-    backgroundColor: colors.gray[50],
-    borderRadius: spacing.spacing3,
-    padding: spacing.spacing4,
-    marginBottom: spacing.spacing4,
-  },
-  applicantHeader: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    marginBottom: spacing.spacing3,
-  },
-  applicantLabel: {
-    fontSize: typography.fontSize.font3,
-    fontWeight: typography.fontWeight.semibold,
-    color: colors.gray[700],
-    marginLeft: spacing.spacing2,
-  },
-  applicantInfo: {
-    flexDirection: 'row',
-    alignItems: 'center',
-  },
-  applicantDetails: {
-    flex: 1,
-  },
-  applicantName: {
-    fontSize: typography.fontSize.font4,
-    fontWeight: typography.fontWeight.semibold,
-    color: colors.gray[900],
-    marginBottom: 2,
-  },
-  applicantSubtext: {
-    fontSize: typography.fontSize.font3,
-    color: colors.gray[600],
-  },
-
-  statusMessageSection: {
-    marginBottom: spacing.spacing4,
-  },
-  statusMessage: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    padding: spacing.spacing3,
-    borderRadius: spacing.spacing2,
-    gap: spacing.spacing2,
-  },
-  statusMessageText: {
-    fontSize: typography.fontSize.font3,
-    fontWeight: typography.fontWeight.medium,
-    flex: 1,
-  },
-
-  actionSection: {
-    flexDirection: 'row',
-    justifyContent: 'flex-end',
-  },
-  viewTeamButton: {
-    backgroundColor: colors.blue[500],
-    flexDirection: 'row',
-    alignItems: 'center',
-    justifyContent: 'center',
-    paddingHorizontal: spacing.spacing5,
-    paddingVertical: spacing.spacing3,
-    borderRadius: spacing.spacing3,
-    gap: spacing.spacing2,
-    shadowColor: colors.blue[500],
-    shadowOffset: { width: 0, height: 2 },
-    shadowOpacity: 0.2,
-    shadowRadius: 4,
-    elevation: 3,
-  },
-  viewTeamButtonText: {
-    color: colors.white,
-    fontSize: typography.fontSize.font3,
-    fontWeight: typography.fontWeight.semibold,
-  },
-  cancelButton: {
-    backgroundColor: colors.red[500],
-    flexDirection: 'row',
-    alignItems: 'center',
-    justifyContent: 'center',
-    paddingHorizontal: spacing.spacing5,
-    paddingVertical: spacing.spacing3,
-    borderRadius: spacing.spacing3,
-    gap: spacing.spacing2,
-    shadowColor: colors.red[500],
-    shadowOffset: { width: 0, height: 2 },
-    shadowOpacity: 0.2,
-    shadowRadius: 4,
-    elevation: 3,
-  },
-  cancelButtonText: {
-    color: colors.white,
-    fontSize: typography.fontSize.font3,
-    fontWeight: typography.fontWeight.semibold,
-  },
-  loadingContainer: {
-    flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-    padding: spacing.spacing5,
-  },
-  loadingText: {
-    marginTop: spacing.spacing3,
-    fontSize: typography.fontSize.font4,
-    color: colors.gray[600],
-  },
-  errorContainer: {
-    flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-    padding: spacing.spacing5,
-  },
-  errorText: {
-    marginTop: spacing.spacing3,
-    fontSize: typography.fontSize.font4,
-    color: colors.gray[600],
-    textAlign: 'center',
-  },
-  retryButton: {
-    marginTop: spacing.spacing4,
-    backgroundColor: colors.blue[500],
-    paddingHorizontal: spacing.spacing5,
-    paddingVertical: spacing.spacing3,
-    borderRadius: spacing.spacing2,
-  },
-  retryButtonText: {
-    color: colors.white,
-    fontSize: typography.fontSize.font3,
-    fontWeight: typography.fontWeight.semibold,
-  },
-  emptyContainer: {
-    flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-    padding: spacing.spacing5,
-  },
-  emptyText: {
-    marginTop: spacing.spacing3,
-    fontSize: typography.fontSize.font4,
-    fontWeight: typography.fontWeight.semibold,
-    color: colors.gray[700],
-  },
-  emptySubtext: {
-    marginTop: spacing.spacing2,
-    fontSize: typography.fontSize.font3,
-    color: colors.gray[500],
-    textAlign: 'center',
-  },
-  footerLoader: {
-    padding: spacing.spacing4,
-    alignItems: 'center',
-  },
-
-  modalContainer: {
-    flex: 1,
-    backgroundColor: colors.white,
-  },
-  modalHeader: {
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    alignItems: 'center',
-    paddingHorizontal: spacing.spacing4,
-    paddingVertical: spacing.spacing4,
-    borderBottomWidth: 1,
-    borderBottomColor: colors.gray[200],
-    backgroundColor: colors.white,
-  },
-  headerContent: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    flex: 1,
-  },
-  iconContainer: {
-    width: 40,
-    height: 40,
-    borderRadius: 20,
-    backgroundColor: colors.orange[50],
-    alignItems: 'center',
-    justifyContent: 'center',
-    marginRight: spacing.spacing3,
-  },
-  modalTitle: {
-    fontSize: typography.fontSize.font5,
-    fontWeight: typography.fontWeight.bold,
-    color: colors.gray[900],
-  },
-  modalContent: {
-    flex: 1,
-    padding: spacing.spacing5,
-  },
-  warningSection: {
-    backgroundColor: colors.orange[50],
-    borderRadius: spacing.spacing3,
-    padding: spacing.spacing4,
-    marginBottom: spacing.spacing5,
-    borderLeftWidth: 4,
-    borderLeftColor: colors.orange[500],
-  },
-  warningText: {
-    fontSize: typography.fontSize.font4,
-    color: colors.orange[700],
-    lineHeight: typography.lineHeight.line6,
-  },
-  teamInfoCard: {
-    backgroundColor: colors.white,
-    borderRadius: spacing.spacing4,
-    padding: spacing.spacing5,
-    marginBottom: spacing.spacing6,
-    borderWidth: 1,
-    borderColor: colors.gray[200],
-    shadowColor: colors.gray[900],
-    shadowOffset: {
-      width: 0,
-      height: 2,
-    },
-    shadowOpacity: 0.05,
-    shadowRadius: 8,
-    elevation: 2,
-  },
-  teamInfoHeader: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    marginBottom: spacing.spacing3,
-  },
-  teamInfoTitle: {
-    fontSize: typography.fontSize.font4,
-    fontWeight: typography.fontWeight.semibold,
-    color: colors.gray[900],
-    marginLeft: spacing.spacing2,
-  },
-  teamInfoContent: {
-    backgroundColor: colors.blue[50],
-    borderRadius: spacing.spacing3,
-    padding: spacing.spacing4,
-  },
-  teamIdLabel: {
-    fontSize: typography.fontSize.font2,
-    color: colors.blue[600],
-    marginBottom: 4,
-    fontWeight: typography.fontWeight.medium,
-  },
-  teamIdValue: {
-    fontSize: typography.fontSize.font5,
-    fontWeight: typography.fontWeight.bold,
-    color: colors.blue[700],
-  },
-  modalButtonContainer: {
-    flexDirection: 'row',
-    gap: spacing.spacing3,
-    padding: spacing.spacing5,
-    paddingTop: spacing.spacing4,
-    paddingBottom: spacing.spacing12,
-    backgroundColor: colors.white,
-    borderTopWidth: 1,
-    borderTopColor: colors.gray[200],
-  },
-  button: {
-    flex: 1,
-    paddingVertical: spacing.spacing4,
-    borderRadius: spacing.spacing4,
-    alignItems: 'center',
-    justifyContent: 'center',
-    flexDirection: 'row',
-    gap: spacing.spacing2,
-  },
-  modalCancelButton: {
-    backgroundColor: colors.gray[100],
-    borderWidth: 1,
-    borderColor: colors.gray[300],
-  },
-  modalCancelButtonText: {
-    fontSize: typography.fontSize.font4,
-    fontWeight: typography.fontWeight.semibold,
-    color: colors.gray[700],
-  },
-  confirmButton: {
-    backgroundColor: colors.red[500],
-    shadowColor: colors.red[500],
-    shadowOffset: {
-      width: 0,
-      height: 4,
-    },
-    shadowOpacity: 0.3,
-    shadowRadius: 8,
-    elevation: 4,
-  },
-  confirmButtonText: {
-    fontSize: typography.fontSize.font4,
-    fontWeight: typography.fontWeight.semibold,
-    color: colors.white,
-  },
-});
