--- src/screens/team/guide/team_guide_styles.ts
@@ -1,216 +1,0 @@
-import { StyleSheet } from 'react-native';
-
-import { theme } from '@/src/theme';
-
-export const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-  },
-  topSection: {
-    flex: 1,
-    paddingHorizontal: 20,
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-  bottomSection: {
-    paddingHorizontal: 20,
-    paddingTop: 20,
-    paddingBottom: 40,
-    justifyContent: 'flex-start',
-  },
-  backButton: {
-    position: 'absolute',
-    top: theme.spacing.spacing17,
-    left: theme.spacing.spacing5,
-    zIndex: 10,
-    backgroundColor: theme.colors.overlay.light,
-    borderRadius: theme.spacing.spacing5,
-    width: theme.spacing.spacing10,
-    height: theme.spacing.spacing10,
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-  subTitle: {
-    fontSize: theme.typography.fontSize.font7,
-    fontWeight: theme.typography.fontWeight.bold,
-    color: theme.colors.text.white,
-    textAlign: 'center',
-    marginTop: theme.spacing.spacing27,
-    marginBottom: theme.spacing.spacing3,
-    opacity: 0.9,
-  },
-  mainTitle: {
-    fontSize: theme.typography.fontSize.font10,
-    fontWeight: theme.typography.fontWeight.bold,
-    color: theme.colors.text.white,
-    textAlign: 'center',
-    lineHeight: theme.typography.lineHeight.line10,
-    marginBottom: theme.spacing.spacing10,
-    marginTop: theme.spacing.spacing4,
-    marginRight: theme.spacing.spacing2,
-  },
-  cardsContainer: {
-    width: '100%',
-    alignItems: 'center',
-    justifyContent: 'center',
-    marginBottom: theme.spacing.spacing10,
-    position: 'relative',
-    height: 380,
-  },
-  card: {
-    backgroundColor: theme.colors.white,
-    borderRadius: theme.spacing.spacing5,
-    padding: theme.spacing.spacing6,
-    shadowColor: theme.colors.black,
-    shadowOffset: { width: 0, height: theme.spacing.spacing2 },
-    shadowOpacity: 0.15,
-    shadowRadius: theme.spacing.spacing4,
-    elevation: 8,
-    flexDirection: 'row',
-    alignItems: 'center',
-    justifyContent: 'space-between',
-    position: 'absolute',
-  },
-  cardFirst: {
-    width: '85%',
-    height: 80,
-    zIndex: 3,
-    transform: [{ rotate: '-3deg' }, { scale: 1.0 }],
-    left: '5%',
-    top: 0,
-    shadowColor: theme.colors.black,
-    shadowOffset: { width: 0, height: theme.spacing.spacing3 },
-    shadowOpacity: 0.2,
-    shadowRadius: theme.spacing.spacing5,
-    elevation: 12,
-  },
-  cardSecond: {
-    width: '82%',
-    height: 75,
-    zIndex: 2,
-    transform: [{ rotate: '3deg' }, { scale: 0.95 }],
-    right: '5%',
-    top: 75,
-    opacity: 0.9,
-    shadowColor: theme.colors.black,
-    shadowOffset: { width: 0, height: theme.spacing.spacing2 },
-    shadowOpacity: 0.15,
-    shadowRadius: theme.spacing.spacing4,
-    elevation: 8,
-  },
-  cardThird: {
-    width: '78%',
-    height: 70,
-    zIndex: 1,
-    transform: [{ rotate: '-2deg' }, { scale: 0.9 }],
-    left: '8%',
-    top: 150,
-    opacity: 0.8,
-    shadowColor: theme.colors.black,
-    shadowOffset: { width: 0, height: theme.spacing.spacing1 },
-    shadowOpacity: 0.1,
-    shadowRadius: theme.spacing.spacing3,
-    elevation: 6,
-  },
-  cardText: {
-    fontSize: theme.typography.fontSize.font3,
-    fontWeight: theme.typography.fontWeight.semibold,
-    color: theme.colors.text.main,
-    flex: 1,
-    lineHeight: theme.typography.lineHeight.line4,
-    flexShrink: 1,
-  },
-  cardButton: {
-    width: 14,
-    height: 14,
-    borderRadius: 7,
-    marginLeft: theme.spacing.spacing3,
-    shadowColor: theme.colors.black,
-    shadowOffset: { width: 0, height: theme.spacing.spacing1 },
-    shadowOpacity: 0.2,
-    shadowRadius: theme.spacing.spacing1,
-    elevation: 3,
-  },
-  cardPressed: {
-    transform: [{ scale: 0.98 }],
-    opacity: 0.8,
-  },
-  instructionText: {
-    fontSize: theme.typography.fontSize.font5,
-    fontWeight: theme.typography.fontWeight.semibold,
-    color: theme.colors.text.white,
-    textAlign: 'center',
-    marginBottom: theme.spacing.spacing5,
-    lineHeight: theme.typography.lineHeight.line7,
-  },
-  buttonContainer: {
-    position: 'absolute',
-    bottom: theme.spacing.spacing10,
-    left: theme.spacing.spacing5,
-    right: theme.spacing.spacing5,
-  },
-  createButton: {
-    backgroundColor: theme.colors.blue[500],
-    borderRadius: theme.spacing.spacing3,
-    paddingVertical: 14,
-    paddingHorizontal: theme.spacing.spacing6,
-    flexDirection: 'row',
-    marginBottom: theme.spacing.spacing3,
-    justifyContent: 'center',
-    alignItems: 'center',
-    shadowColor: theme.colors.blue[500],
-    shadowOffset: { width: 0, height: theme.spacing.spacing1 },
-    shadowOpacity: 0.3,
-    shadowRadius: theme.spacing.spacing2,
-    elevation: 4,
-  },
-  createButtonText: {
-    fontSize: theme.typography.fontSize.font5,
-    fontWeight: theme.typography.fontWeight.bold,
-    color: theme.colors.text.white,
-    marginLeft: theme.spacing.spacing2,
-    flexShrink: 1,
-  },
-  joinButton: {
-    backgroundColor: theme.colors.white,
-    borderWidth: 2,
-    borderColor: theme.colors.blue[500],
-    borderRadius: theme.spacing.spacing3,
-    paddingVertical: theme.spacing.spacing3,
-    paddingHorizontal: theme.spacing.spacing5,
-    flexDirection: 'row',
-    justifyContent: 'center',
-    alignItems: 'center',
-    shadowColor: theme.colors.black,
-    shadowOffset: { width: 0, height: theme.spacing.spacing1 },
-    shadowOpacity: 0.1,
-    shadowRadius: theme.spacing.spacing1,
-    elevation: 2,
-    marginBottom: theme.spacing.spacing2,
-  },
-  joinButtonText: {
-    fontSize: theme.typography.fontSize.font5,
-    fontWeight: theme.typography.fontWeight.bold,
-    color: theme.colors.blue[500],
-    marginLeft: theme.spacing.spacing2,
-    flexShrink: 1,
-  },
-  notificationBadge: {
-    position: 'absolute',
-    top: -8,
-    right: -8,
-    backgroundColor: theme.colors.red[500],
-    borderRadius: 10,
-    minWidth: 20,
-    height: 20,
-    justifyContent: 'center',
-    alignItems: 'center',
-    borderWidth: 2,
-    borderColor: theme.colors.white,
-  },
-  notificationText: {
-    fontSize: 12,
-    fontWeight: theme.typography.fontWeight.bold,
-    color: theme.colors.text.white,
-  },
-});
