--- src/screens/team/join/university_team_list_style.ts
@@ -1,363 +1,0 @@
-import { StyleSheet } from 'react-native';
-
-import { theme } from '@/src/theme';
-
-export const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: theme.colors.background.main,
-  },
-  headerSection: {
-    paddingHorizontal: theme.spacing.spacing5,
-    paddingTop: theme.spacing.spacing4,
-    paddingBottom: theme.spacing.spacing3,
-  },
-  filterRow: {
-    flexDirection: 'row',
-    justifyContent: 'flex-end',
-    alignItems: 'center',
-    marginTop: theme.spacing.spacing2,
-  },
-  title: {
-    fontSize: theme.typography.fontSize.font7,
-    fontWeight: theme.typography.fontWeight.bold,
-    color: theme.colors.text.main,
-    marginBottom: theme.spacing.spacing2,
-    textAlign: 'center',
-  },
-  subtitle: {
-    fontSize: theme.typography.fontSize.font4,
-    color: theme.colors.text.sub,
-    textAlign: 'center',
-    lineHeight: theme.typography.lineHeight.line6,
-    marginBottom: theme.spacing.spacing3,
-  },
-  loadingContainer: {
-    flex: 1,
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-  loadingText: {
-    fontSize: theme.typography.fontSize.font4,
-    color: theme.colors.text.sub,
-  },
-  teamList: {
-    paddingHorizontal: theme.spacing.spacing5,
-    paddingBottom: theme.spacing.spacing10,
-  },
-  teamCard: {
-    backgroundColor: theme.colors.background.sub,
-    borderRadius: theme.spacing.spacing3,
-    padding: theme.spacing.spacing4,
-    marginBottom: theme.spacing.spacing3,
-    borderWidth: 1,
-    borderColor: theme.colors.border.light,
-  },
-  teamHeader: {
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    alignItems: 'flex-start',
-    marginBottom: theme.spacing.spacing2,
-  },
-  teamName: {
-    fontSize: theme.typography.fontSize.font5,
-    fontWeight: theme.typography.fontWeight.semibold,
-    color: theme.colors.text.main,
-    marginBottom: theme.spacing.spacing1,
-  },
-  teamTypeBadge: {
-    backgroundColor: theme.colors.blue[50],
-    paddingHorizontal: theme.spacing.spacing3,
-    paddingVertical: theme.spacing.spacing1,
-    borderRadius: theme.spacing.spacing2,
-  },
-  teamTypeText: {
-    fontSize: theme.typography.fontSize.font2,
-    color: theme.colors.blue[500],
-    fontWeight: theme.typography.fontWeight.medium,
-  },
-  teamDescription: {
-    fontSize: theme.typography.fontSize.font3,
-    color: theme.colors.text.sub,
-    lineHeight: theme.typography.lineHeight.line5,
-    marginBottom: theme.spacing.spacing3,
-  },
-  teamInfo: {
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    marginBottom: theme.spacing.spacing3,
-    paddingHorizontal: theme.spacing.spacing1,
-  },
-  infoItem: {
-    flex: 1,
-    alignItems: 'center',
-    paddingVertical: theme.spacing.spacing1,
-  },
-  infoLabel: {
-    fontSize: theme.typography.fontSize.font2,
-    color: theme.colors.text.light,
-    marginBottom: theme.spacing.spacing1,
-    fontWeight: theme.typography.fontWeight.medium,
-  },
-  infoValue: {
-    fontSize: theme.typography.fontSize.font3,
-    color: theme.colors.text.main,
-    fontWeight: theme.typography.fontWeight.semibold,
-  },
-  joinButton: {
-    backgroundColor: theme.colors.blue[500],
-    borderRadius: theme.spacing.spacing2,
-    paddingVertical: theme.spacing.spacing2,
-    alignItems: 'center',
-    opacity: 0.8,
-  },
-  joinButtonText: {
-    fontSize: theme.typography.fontSize.font3,
-    fontWeight: theme.typography.fontWeight.semibold,
-    color: theme.colors.white,
-  },
-  filterButton: {
-    backgroundColor: theme.colors.blue[500],
-    width: 40,
-    height: 40,
-    borderRadius: 20,
-    alignItems: 'center',
-    opacity: 0.9,
-    justifyContent: 'center',
-    shadowColor: theme.colors.blue[500],
-    shadowOffset: {
-      width: 0,
-      height: 2,
-    },
-    shadowOpacity: 0.2,
-    shadowRadius: 4,
-    elevation: 3,
-  },
-  modalOverlay: {
-    flex: 1,
-    backgroundColor: 'rgba(0, 0, 0, 0.5)',
-    justifyContent: 'flex-end',
-  },
-  modalBackdrop: {
-    flex: 1,
-  },
-  filterModal: {
-    backgroundColor: theme.colors.background.main,
-    borderTopLeftRadius: theme.spacing.spacing5,
-    borderTopRightRadius: theme.spacing.spacing5,
-    maxHeight: '75%',
-    minHeight: '60%',
-    shadowColor: theme.colors.black,
-    shadowOffset: {
-      width: 0,
-      height: -4,
-    },
-    shadowOpacity: 0.1,
-    shadowRadius: 12,
-    elevation: 8,
-  },
-  filterHeader: {
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    alignItems: 'center',
-    paddingHorizontal: theme.spacing.spacing5,
-    paddingTop: theme.spacing.spacing4,
-    paddingBottom: theme.spacing.spacing4,
-    borderBottomWidth: 1,
-    borderBottomColor: theme.colors.border.light,
-  },
-  dragHandle: {
-    position: 'absolute',
-    top: theme.spacing.spacing2,
-    left: '50%',
-    marginLeft: -20,
-    width: 40,
-    height: 4,
-    backgroundColor: theme.colors.border.light,
-    borderRadius: theme.spacing.spacing1,
-  },
-  filterTitle: {
-    fontSize: theme.typography.fontSize.font6,
-    fontWeight: theme.typography.fontWeight.bold,
-    color: theme.colors.text.main,
-  },
-  closeButton: {
-    fontSize: theme.typography.fontSize.font5,
-    color: theme.colors.text.sub,
-    padding: theme.spacing.spacing1,
-  },
-  filterContent: {
-    paddingHorizontal: theme.spacing.spacing5,
-    paddingVertical: theme.spacing.spacing5,
-    flex: 1,
-  },
-  filterSection: {
-    marginBottom: theme.spacing.spacing6,
-    paddingVertical: theme.spacing.spacing2,
-  },
-  filterSectionTitle: {
-    fontSize: theme.typography.fontSize.font4,
-    fontWeight: theme.typography.fontWeight.semibold,
-    color: theme.colors.text.main,
-    marginBottom: theme.spacing.spacing4,
-  },
-  filterOptions: {
-    flexDirection: 'row',
-    flexWrap: 'wrap',
-    gap: theme.spacing.spacing2,
-  },
-  filterOption: {
-    paddingHorizontal: theme.spacing.spacing4,
-    paddingVertical: theme.spacing.spacing3,
-    borderRadius: theme.spacing.spacing3,
-    borderWidth: 1.5,
-    borderColor: theme.colors.border.light,
-    backgroundColor: theme.colors.background.sub,
-    minWidth: 90,
-    alignItems: 'center',
-    shadowColor: theme.colors.black,
-    shadowOffset: {
-      width: 0,
-      height: 1,
-    },
-    shadowOpacity: 0.05,
-    shadowRadius: 2,
-    elevation: 1,
-  },
-  filterOptionSelected: {
-    backgroundColor: theme.colors.blue[500],
-    borderColor: theme.colors.blue[500],
-    shadowColor: theme.colors.blue[500],
-    shadowOffset: {
-      width: 0,
-      height: 2,
-    },
-    shadowOpacity: 0.2,
-    shadowRadius: 4,
-    elevation: 3,
-  },
-  filterOptionText: {
-    fontSize: theme.typography.fontSize.font3,
-    color: theme.colors.text.main,
-    fontWeight: theme.typography.fontWeight.medium,
-  },
-  filterOptionTextSelected: {
-    color: theme.colors.white,
-    fontWeight: theme.typography.fontWeight.semibold,
-  },
-  sliderSection: {
-    alignItems: 'center',
-    marginTop: theme.spacing.spacing3,
-    paddingVertical: theme.spacing.spacing2,
-  },
-  memberCountSlider: {
-    width: '100%',
-    marginBottom: theme.spacing.spacing3,
-  },
-  sliderTrack: {
-    height: 4,
-    backgroundColor: theme.colors.border.light,
-    borderRadius: theme.spacing.spacing1,
-    position: 'relative',
-    marginBottom: theme.spacing.spacing3,
-  },
-  sliderTrackBackground: {
-    position: 'absolute',
-    top: 0,
-    left: 0,
-    right: 0,
-    height: 4,
-    backgroundColor: theme.colors.border.light,
-    borderRadius: theme.spacing.spacing1,
-  },
-  sliderActiveTrack: {
-    position: 'absolute',
-    top: 0,
-    height: 4,
-    backgroundColor: theme.colors.blue[500],
-    borderRadius: theme.spacing.spacing1,
-  },
-  sliderLabels: {
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-  },
-  sliderLabel: {
-    paddingHorizontal: theme.spacing.spacing1,
-    paddingVertical: theme.spacing.spacing1,
-    borderRadius: theme.spacing.spacing1,
-  },
-  sliderLabelActive: {
-    backgroundColor: theme.colors.blue[50],
-  },
-  sliderLabelText: {
-    fontSize: theme.typography.fontSize.font1,
-    color: theme.colors.text.sub,
-    textAlign: 'center',
-  },
-  sliderLabelTextActive: {
-    color: theme.colors.blue[500],
-    fontWeight: theme.typography.fontWeight.medium,
-  },
-
-  filterActions: {
-    flexDirection: 'row',
-    paddingHorizontal: theme.spacing.spacing5,
-    paddingTop: theme.spacing.spacing2,
-    paddingBottom: theme.spacing.spacing6,
-    borderTopWidth: 1,
-    borderTopColor: theme.colors.border.light,
-    gap: theme.spacing.spacing3,
-    backgroundColor: theme.colors.background.main,
-  },
-  resetButton: {
-    flex: 1,
-    backgroundColor: theme.colors.background.sub,
-    paddingVertical: theme.spacing.spacing4,
-    borderRadius: theme.spacing.spacing3,
-    alignItems: 'center',
-    borderWidth: 1.5,
-    borderColor: theme.colors.border.light,
-    shadowColor: theme.colors.black,
-    shadowOffset: {
-      width: 0,
-      height: 1,
-    },
-    shadowOpacity: 0.05,
-    shadowRadius: 2,
-    elevation: 1,
-  },
-  resetButtonText: {
-    fontSize: theme.typography.fontSize.font4,
-    fontWeight: theme.typography.fontWeight.medium,
-    color: theme.colors.text.main,
-  },
-  applyButton: {
-    flex: 1,
-    backgroundColor: theme.colors.blue[500],
-    paddingVertical: theme.spacing.spacing4,
-    borderRadius: theme.spacing.spacing3,
-    alignItems: 'center',
-    shadowColor: theme.colors.blue[500],
-    shadowOffset: {
-      width: 0,
-      height: 2,
-    },
-    shadowOpacity: 0.2,
-    shadowRadius: 4,
-    elevation: 3,
-  },
-  applyButtonText: {
-    fontSize: theme.typography.fontSize.font4,
-    fontWeight: theme.typography.fontWeight.semibold,
-    color: theme.colors.white,
-  },
-  teamTitleSection: {
-    flex: 1,
-    marginRight: theme.spacing.spacing3,
-  },
-  universityName: {
-    fontSize: theme.typography.fontSize.font3,
-    color: theme.colors.text.sub,
-    fontWeight: theme.typography.fontWeight.medium,
-  },
-});
