+++ src/screens/team/management/edit_screen.tsx
@@ -15,7 +15,7 @@ import InfoSection from '@/src/components/team/sections/info_section';
 import { CustomHeader } from '@/src/components/ui/custom_header';
 import GlobalErrorFallback from '@/src/components/ui/global_error_fallback';
 import { useTeam, useUpdateTeamMutation } from '@/src/hooks/queries';
-import { styles } from '@/src/screens/team/management/team_edit_styles';
+import { styles } from '@/src/screens/team/management/edit_styles';
 import { colors } from '@/src/theme';
 import {
   DEFAULT_SKILL_LEVEL,
@@ -25,11 +25,11 @@ import {
 } from '@/src/types/team';
 import type { SkillLevel, TeamType } from '@/src/types/team';
 
-interface TeamEditScreenProps {
+interface EditScreenProps {
   teamId: string | number;
 }
 
-export default function TeamEditScreen({ teamId }: TeamEditScreenProps) {
+export default function EditScreen({ teamId }: EditScreenProps) {
   const [formData, setFormData] = useState({
     name: '',
     description: '',
