+++ src/screens/team/management/management_styles.ts
